pyproject.toml
[project]
name = "trading-agent"
version = "1.0.0"
description = "AI-powered Trading Analysis Agent — LangGraph + Ollama + Polygon + PostgreSQL"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",
    "httpx>=0.27.0",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "pgvector>=0.3.0",
    "psycopg2-binary>=2.9.0",
    "polygon-api-client>=1.14.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=24.1.0",
    "python-docx>=1.1.0",
    "orjson>=3.10.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "rich>=13.9.0",
    "apscheduler>=3.10.0",
    "websockets>=13.0",
]

[tool.setuptools]
packages = ["app", "data", "config"]


[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
